Changing uint to int.

+++ thrii/Drawer.cs
@@ -8,12 +8,12 @@ namespace thrii
 	{
 		public RenderWindow Window { get; set; }
 
-		public Drawer(uint width, uint height, string name, string iconPath)
+		public Drawer(int width, int height, string name, string iconPath)
 		{
 			var windowSettings = new ContextSettings();
 			windowSettings.AntialiasingLevel = 8;
 
-			Window = new RenderWindow(new VideoMode(width, height), name, Styles.Close, windowSettings);
+			Window = new RenderWindow(new VideoMode((uint)width, (uint)height), name, Styles.Close, windowSettings);
 			SetIcon(iconPath);
 			Window.Closed += OnClosed;
 		}
